wip(conftest): reduce test code redundancy by moving one GeneratorAgent into conftest.py
