+++ tests/test_send_plot_image.py
@@ -1,42 +1,8 @@
 import time
 
-import matplotlib
-import matplotlib.pyplot as plt
-import numpy as np
-
-from agentMET4FOF.agents.base_agents import AgentMET4FOF, AgentNetwork, MonitorAgent
-
-matplotlib.use("Agg")
-
-
-class GeneratorAgent(AgentMET4FOF):
-    def create_graph(self):
-        # Fixing random state for reproducibility
-        np.random.seed(19680801)
-
-        dt = 0.01
-        t = np.arange(0, 30, dt)
-        nse1 = np.random.randn(len(t))  # white noise 1
-        nse2 = np.random.randn(len(t))  # white noise 2
-
-        # Two signals with a coherent part at 10Hz and a random part
-        s1 = np.sin(2 * np.pi * 10 * t) + nse1
-        s2 = np.sin(2 * np.pi * 10 * t) + nse2
-
-        fig, axs = plt.subplots(2, 1)
-        axs[0].plot(t, s1, t, s2)
-        axs[0].set_xlim(0, 2)
-        axs[0].set_xlabel("time")
-        axs[0].set_ylabel("s1 and s2")
-        axs[0].grid(True)
-
-        cxy, f = axs[1].cohere(s1, s2, 256, 1.0 / dt)
-        axs[1].set_ylabel("coherence")
-
-        return fig
-
-    def dummy_send_graph(self, mode="image"):
-        self.send_plot(self.create_graph(), mode=mode)
+from agentMET4FOF.agents import AgentNetwork
+from agentMET4FOF.agents.base_agents import MonitorAgent
+from tests.conftest import GeneratorAgent
 
 
 def test_send_plot():
